requirements.txt
openai
tiktoken
Flask
python-dotenv
python-dateutil
requests
pandas
pytz
tqdm
beautifulsoup4
feedgen
feedparser